Remove jQuery event namespaces.
What
=
Remove jQuery event namespaces and replaced with traditional removing of event handlers by referencing the function that was registered.

Why
=
This is being done to step away from jQuery and move to plain JavaScript. In plain JavaScript namespaces don't exist and instead to remove an event handler registration a reference to the function must be passed.