Replaced fallback to `$(window).height()` with plain a JavaScript version.
What
=
Replaced fallback to `$(window).height()` with plain a JavaScript version that supports IE5+.

Why
=
This is being done to step away from jQuery and move to plain JavaScript. In plain JavaScript `window.innerHeight` is supported by modern browsers and IE9+, the other fallbacks are for <IE9 back to IE5.

Reference
=
http://www.w3schools.com/js/js_window.asp