Replaced `$(window).scrollTop()` with plain a JavaScript version.
What
=
Replaced fallback to `$(window).scrollTop()` with plain a JavaScript version that gets the page y-axis offset due to scrolling.

Why
=
This is being done to step away from jQuery and move to plain JavaScript. In plain JavaScript `window.pageYOffset` is supported by modern browsers and IE9+, the other fallbacks are for <IE9.