Replaced `$(window).on(...)` and `$(window).off(...)` with plain a JavaScript version.
What
=
Replaced `$(window).on(...)` and `$(window).off(...)` with plain JavaScript `addEventListener` and `removeEventListener` and alternatives for <IE9.

Why
=
This is being done to step away from jQuery and move to plain JavaScript. In plain JavaScript `window.addEventListener` is supported by modern browsers and IE9+ and the use of `attachEvent` and others adds support for <IE9.