Replace `$(elem).offset().top` with plain JavaScript.
What
=
Replace `$(elem).offset().top` with plain JavaScript, using `getBoundingClientRect` and the `pageYOffset` to calculate the offset of the element into the document.

Why
=
This is being done to step away from jQuery and move to plain JavaScript. In plain JavaScript `getBoundingClientRect` tells us the offset of the element into the viewport and we can add it's value to the `pageYOffset`, which is the current scroll position, to find out the position of the element in the document.