Allow `elements` to be specified as the `DOMElement`s to allow for increased browser compatibility.
### What
Allow `elements` to be specified as the `DOMElement`s instead of string selectors.

### Why
When the developer specifies `elements` as a string we use `document.querySelector(...)` to get the first `DOMElement` matching the selector. This is only supported in IE8+. To allow support in older browsers we can simply allow the developer to give us the actual `DOMElement` instead in `elements` and they can then use any method at their disposal to lookup the element, like `document.getElementById`.