Try really hard to make DOMElement lookup from `elements` work on all browsers.
What
=
Try really hard to make DOMElement lookup from `elements` work on all browsers by using jQuery if it's available, `document.querySelector` if not, and falling back to `document.getElementById` if needed and the selector is an ID.

Why
=
For maximum backwards compatibility with browsers, and for developers already using the library. If developers using the library are using selectors unsupported by browsers but supported by jQuery their uses will continue to work successfully. Additionally if jQuery isn't included and the library is used on older browsers there are simple ways the developer can choose to use the library to get maximum browser compatibility.